requirements.txt
# Core stack
numpy>=1.24
pandas>=2.0
xarray>=2023.0
rasterio>=1.3
shapely>=2.0
geopandas>=0.14
scikit-learn>=1.3
scipy>=1.10

# ML/AI
torch>=2.1
onnxruntime>=1.16
lightgbm>=4.1
hmmlearn>=0.3

# Data handling
netCDF4>=1.6
h5py>=3.9
pyyaml>=6.0
joblib>=1.3
tqdm>=4.66

# Plotting and EDA
matplotlib>=3.8
seaborn>=0.12

# Web3 + Emitters
web3>=6.12
reportlab>=4.0
PyPDF2>=3.0

# API + CLI
fastapi>=0.105
uvicorn>=0.25
click>=8.1
typer>=0.9

# Testing
pytest-cov>=4.1

tqdm>=4.66
loguru>=0.7


# --- Testing ---
pytest>=7.4
coverage>=7.2
pre-commit>=3.3

# --- Optional: Audio / CLI support ---
click>=8.1
pydub>=0.25
openpyxl>=3.1

# === Transformers / NLP / Tokenization ===
transformers>=4.42
sentencepiece>=0.2.0
tokenizers>=0.19.1
datasets>=2.19

# === Vision / Multimodal ===
torchvision==0.16.0
torchaudio==2.1.0
opencv-python>=4.9
Pillow>=10.0

# === HuggingFace tooling ===
huggingface-hub>=0.22.2
safetensors>=0.4.2

# === Probabilistic Programming / Bayesian Modeling ===
pyro-ppl==1.9.1           # Compatible with torch==2.1.0, includes MCMC + SVI inference

# === Jupyter / Notebook Kernel Support ===
ipykernel==6.29.4         # Stable kernel for Python 3.11

# === Development Tooling ===
black==24.4.2             # Code formatter
mypy==1.10.0              # Static type checker
isort==5.13.2             # Import sorting
flake8==6.1.0             # Linting
pre-commit==3.3.0         # Git hooks for code quality
pylint==3.1.1           # Code analysis